lab1/openai-photo-demo/requirements.txt
openai
python-dotenv
elevenlabs
huggingface_hub
